On install, skip symlinks that have been created.
Before it would try to replace them; in the case of directories this would
fail to replace the link and instead create a new link inside the symlinked
dir.  It is probably best to notify the user and allow the user to manually
remove them first.  Make uninstall can also be used.